video_core: Move HardwareVertex to RasterizerAccelerated
